twelvelabs/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.0.3
Flask-CORS==5.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
twelvelabs==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
